fixed migration to spring boot 2, plus now added properties in articles
