+++ spring-security-mvc-session/src/main/java/com/baeldung/SpringSessionApplication.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
