+++ spring-security-mvc-session/src/main/java/com/baeldung/security/MySimpleUrlAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package org.baeldung.security;
+package com.baeldung.security;
 
 import java.io.IOException;
 import java.util.Collection;
