+++ spring-security-mvc-session/src/main/java/com/baeldung/security/SessionFilter.java
@@ -1,4 +1,4 @@
-package org.baeldung.security;
+package com.baeldung.security;
 
 import java.io.IOException;
 import java.util.Arrays;
