+++ spring-security-mvc-session/src/main/java/com/baeldung/web/SessionListenerWithMetrics.java
@@ -1,12 +1,11 @@
-package org.baeldung.web;
+package com.baeldung.web;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
 import javax.servlet.http.HttpSessionEvent;
 import javax.servlet.http.HttpSessionListener;
 
-import org.baeldung.monitoring.MetricRegistrySingleton;
-
+import com.baeldung.monitoring.MetricRegistrySingleton;
 import com.codahale.metrics.Counter;
 
 public class SessionListenerWithMetrics implements HttpSessionListener {
