--- spring-security-mvc-session/src/main/java/org/baeldung/spring/MvcConfig.java
@@ -1,44 +1,0 @@
-package org.baeldung.spring;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.ViewResolver;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-import org.springframework.web.servlet.view.JstlView;
-
-@EnableWebMvc
-@Configuration
-public class MvcConfig implements WebMvcConfigurer {
-
-    public MvcConfig() {
-        super();
-    }
-
-    // API
-
-    @Override
-    public void addViewControllers(final ViewControllerRegistry registry) {
-
-        registry.addViewController("/anonymous.html");
-
-        registry.addViewController("/login.html");
-        registry.addViewController("/homepage.html");
-        registry.addViewController("/sessionExpired.html");
-        registry.addViewController("/invalidExpired.html");
-        registry.addViewController("/console.html");
-    }
-
-    @Bean
-    public ViewResolver viewResolver() {
-        final InternalResourceViewResolver bean = new InternalResourceViewResolver();
-
-        bean.setViewClass(JstlView.class);
-        bean.setPrefix("/WEB-INF/view/");
-        bean.setSuffix(".jsp");
-
-        return bean;
-    }
-}
