+++ spring-security-mvc-session/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+server.servlet.session.timeout=60s
+
+spring.mvc.view.prefix=/WEB-INF/view/
+spring.mvc.view.suffix=.jsp
+
+## Secure Session Cookie configurations
+#server.servlet.session.cookie.http-only=true
+#server.servlet.session.cookie.secure=true
