--- spring-security-mvc-session/src/test/java/org/baeldung/SpringContextIntegrationTest.java
@@ -1,19 +1,0 @@
-package org.baeldung;
-
-import org.baeldung.spring.MvcConfig;
-import org.baeldung.spring.SecSecurityConfig;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = { MvcConfig.class, SecSecurityConfig.class })
-@WebAppConfiguration
-public class SpringContextIntegrationTest {
-
-    @Test
-    public void whenSpringContextIsBootstrapped_thenNoExceptions() {
-    }
-}
